+++ Merger/PathParameterMerger.php
@@ -30,40 +30,31 @@ class PathParameterMerger
     /**
      * @var MergeStrategyInterface
      */
-    private $defaultMergeStrategy;
+    private $mergeStrategy;
 
     /**
      * @param MergeStrategyInterface $defaultMergeStrategy
      */
     public function __construct(MergeStrategyInterface $defaultMergeStrategy)
     {
-        $this->defaultMergeStrategy = $defaultMergeStrategy;
+        $this->mergeStrategy = $defaultMergeStrategy;
     }
 
     /**
      * @param Path $swaggerPath
      * @param string $requestMethod
      * @param Definitions $definitions
-     * @param MergeStrategyInterface|null $mergeStrategy
      *
      * @return Schema
      */
-    public function merge(
-        Path $swaggerPath,
-        string $requestMethod,
-        Definitions $definitions,
-        ?MergeStrategyInterface $mergeStrategy = null
-    ): Schema {
-        if (!$mergeStrategy) {
-            $mergeStrategy = $this->defaultMergeStrategy;
-        }
-
+    public function merge(Path $swaggerPath, string $requestMethod, Definitions $definitions): Schema
+    {
         $swaggerOperation = $swaggerPath->getOperation($requestMethod);
 
         /** @var Parameter $parameter */
         foreach ($swaggerOperation->getParameters() as $parameter) {
             if ($parameter->getIn() !== ParameterLocationEnum::IN_BODY) {
-                $mergeStrategy->addParameter(
+                $this->mergeStrategy->addParameter(
                     $parameter->getIn(),
                     $parameter->getName(),
                     $parameter->toArray() + ['title' => $parameter->getIn()],
@@ -87,7 +78,7 @@ class PathParameterMerger
                 $required = array_flip($required);
 
                 foreach ($refDefinition->getProperties() as $defName => $defItem) {
-                    $mergeStrategy->addParameter(
+                    $this->mergeStrategy->addParameter(
                         $parameter->getIn(),
                         $defName,
                         $defItem->toArray() + ['title' => $parameter->getIn()],
@@ -104,7 +95,7 @@ class PathParameterMerger
                 $required = array_flip($required);
 
                 foreach ($parameterSchema->getProperties() as $bodyItemName => $currentBodyItem) {
-                    $mergeStrategy->addParameter(
+                    $this->mergeStrategy->addParameter(
                         $parameter->getIn(),
                         $bodyItemName,
                         $currentBodyItem->toArray() + ['title' => $parameter->getIn()],
@@ -116,7 +107,7 @@ class PathParameterMerger
             }
 
             // body as scalar
-            $mergeStrategy->addParameter(
+            $this->mergeStrategy->addParameter(
                 $parameter->getIn(),
                 $parameter->getName(),
                 $parameterSchema->toArray() + ['title' => $parameter->getIn()],
@@ -127,11 +118,11 @@ class PathParameterMerger
         $mergedSchema = new Schema();
         $mergedSchema->merge([
             'type' => 'object',
-            'properties' => $mergeStrategy->getParameters(),
-            'required' => $mergeStrategy->getRequired(),
+            'properties' => $this->mergeStrategy->getParameters(),
+            'required' => $this->mergeStrategy->getRequired(),
         ]);
 
-        $mergeStrategy->clean();
+        $this->mergeStrategy->clean();
 
         return $mergedSchema;
     }